Updates light packets
